splited raytracer class to raytracer and renderer + namespace
